# server/athenian/api/openapi/openapi.yaml
openapi: 3.0.0
info:
  license:
    name: CC-BY-4.0
  title: Athenian Dashboard API
  version: 1.0.0
servers:
- url: https://api.owl.athenian.co/v1
paths:
...
  /user:
    get:
      operationId: get_user
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: Details about the current user.
      security:
      - bearerAuth: []
      summary: Show details about the current user.
      x-openapi-router-controller: athenian.api.controllers.user_controller
      x-codegen-request-body-name: body
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: athenian.api.controllers.security_controller.info_from_bearerAuth
  schemas:
...
    User:
      description: User details. "updated" and "accounts" are populated only for the
        current user.
      example:
        id: github|60340680
        native_id: 60340680
        name: Groundskeeper Willie
        email: bot@athenian.co
        picture: https://avatars0.githubusercontent.com/u/60340680?v=4
        updated: 2020-01-23T12:00:00Z
      properties:
        id:
          description: Auth0 user identifier.
          type: string
        native_id:
          description: Auth backend user identifier.
          type: string
        name:
          description: Full name of the user.
          type: string
        email:
          description: Email of the user.
          type: string
        picture:
          description: Avatar URL of the user.
          type: string
        updated:
          description: Date and time of the last profile update.
          format: date-time
          type: string
        accounts:
          description: Mapping between account IDs the user is a member of and is_admin
            flags.
          example:
            1: true
            2: false
          type: object
      required:
      - email
      - id
      - name
      - native_id
      - picture
      type: object
...